Split App.tsx into shell and views, and styles.css into per-room files
App.tsx keeps routing, data loading, the event stream, and shortcuts.
Chrome (sidebar, top bar, mobile nav, palette, access screens, shared
primitives, nav and theme registries) lives in src/shell/. The rooms that
were inline (Live, Overview, Sessions, Activity, Library, Memory, Themes)
join the others in src/views/. Number and time formatters move to
src/format.ts with tests and replace three private copies.

styles.css is partitioned in source order into src/styles/*.css behind
src/styles/index.css. The built CSS is byte-identical.
