+++ src/main.tsx
@@ -2,9 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import '@fontsource-variable/syne'
 import '@fontsource/ibm-plex-mono/400.css'
-import './styles.css'
-import './styles/fleet.css'
-import './styles/minimal-calm.css'
+import './styles/index.css'
 import App from './App'
 import { ErrorBoundary } from './ErrorBoundary'
