+++ src/styles/access.css
@@ -1,0 +1,120 @@
+/* Access stage: boot and sign-in cards. */
+
+/* Authentication gate */
+.access-screen {
+  position: relative;
+  min-height: 100vh;
+  display: grid;
+  place-items: center;
+  overflow: hidden;
+  background:
+    linear-gradient(90deg, rgba(255,255,255,.018) 1px, transparent 1px) 0 0 / 72px 72px,
+    linear-gradient(rgba(255,255,255,.018) 1px, transparent 1px) 0 0 / 72px 72px,
+    var(--ink);
+}
+
+.access-card {
+  position: relative;
+  z-index: 2;
+  width: min(460px, calc(100vw - 32px));
+  padding: 34px;
+  border: 1px solid var(--line-strong);
+  background: rgba(11,12,10,.93);
+  box-shadow: 0 30px 90px rgba(0,0,0,.55);
+}
+
+.access-card::after {
+  content: '';
+  position: absolute;
+  width: 85px;
+  height: 85px;
+  right: -1px;
+  top: -1px;
+  border-top: 1px solid var(--lime);
+  border-right: 1px solid var(--lime);
+  clip-path: polygon(42% 0, 100% 0, 100% 58%);
+}
+
+.access-brand { padding: 0 0 32px; }
+
+.access-card h1 {
+  margin: 22px 0 12px;
+  font-family: 'Syne Variable', sans-serif;
+  font-size: clamp(34px, 8vw, 54px);
+  line-height: .88;
+  letter-spacing: -.055em;
+}
+
+.access-card h1 em {
+  color: var(--muted);
+  font-style: normal;
+  font-weight: 420;
+}
+
+.access-card > p {
+  margin: 0 0 24px;
+  color: var(--muted);
+  font-size: var(--type-body);
+  line-height: 1.75;
+}
+
+.access-card > label { display: grid; gap: 8px; }
+
+.access-card > label > span {
+  color: var(--muted);
+  font-size: var(--type-meta);
+  letter-spacing: .15em;
+}
+
+.access-error {
+  margin-top: 10px;
+  color: #ff9f91;
+  font-size: var(--type-body);
+}
+
+.boot-card {
+  min-height: 220px;
+  display: grid;
+  place-content: center;
+  justify-items: center;
+  gap: 22px;
+}
+
+.boot-card .brand-logo-shell {
+  width: 58px;
+  height: 58px;
+}
+
+.boot-card .brand-logo {
+  width: 60px;
+  height: 60px;
+}
+
+.boot-card strong {
+  color: var(--muted);
+  font-size: var(--type-meta);
+  letter-spacing: .17em;
+}
+
+.loading-bars { display: flex; gap: 4px; }
+
+.loading-bars i {
+  width: 22px;
+  height: 2px;
+  background: var(--line-strong);
+  animation: bootBars 1.5s ease-in-out infinite;
+}
+
+.loading-bars i:nth-child(2) { animation-delay: .12s; }
+.loading-bars i:nth-child(3) { animation-delay: .24s; }
+.loading-bars i:nth-child(4) { animation-delay: .36s; }
+
+@keyframes bootBars {
+  0%, 100% { background: var(--line-strong); }
+  50% { background: var(--lime); box-shadow: 0 0 8px rgba(217,255,67,.5); }
+}
+
+.mobile-bottom-nav,
+.nav-scrim {
+  display: none;
+}
