+++ src/styles/base.css
@@ -1,0 +1,158 @@
+/* Design tokens, resets, scrollbars, and the fatal-error screen. */
+
+:root {
+  font-family: 'IBM Plex Mono', monospace;
+  color: #eeeee8;
+  background: #090a08;
+  font-synthesis: none;
+  text-rendering: optimizeLegibility;
+  --ink: #090a08;
+  --surface: #0e100d;
+  --surface-2: #141612;
+  --surface-3: #1a1d17;
+  --paper: #eeeee8;
+  --muted: #a1a69a;
+  --muted-dim: #777d72;
+  --line: rgba(238, 238, 232, 0.13);
+  --line-strong: rgba(238, 238, 232, 0.28);
+  --lime: #d9ff43;
+  --lime-soft: rgba(217, 255, 67, 0.12);
+  --coral: #ff654f;
+  --coral-soft: rgba(255, 101, 79, 0.12);
+  --amber: #ffca57;
+  --type-micro: 8px;
+  --type-label: 9px;
+  --type-meta: 10px;
+  --type-body: 11px;
+  --type-control: 12px;
+  --sidebar: 236px;
+  --topbar: 64px;
+  --mobile-nav-height: 68px;
+}
+
+* {
+  box-sizing: border-box;
+}
+
+.sr-only {
+  position: absolute;
+  width: 1px;
+  height: 1px;
+  padding: 0;
+  overflow: hidden;
+  clip: rect(0, 0, 0, 0);
+  white-space: nowrap;
+  border: 0;
+}
+
+html {
+  min-width: 320px;
+  min-height: 100%;
+  background: var(--ink);
+}
+
+body {
+  margin: 0;
+  min-width: 320px;
+  min-height: 100vh;
+  overflow-x: hidden;
+}
+
+button,
+input,
+textarea,
+select {
+  font: inherit;
+}
+
+button {
+  color: inherit;
+}
+
+button,
+a {
+  -webkit-tap-highlight-color: transparent;
+  touch-action: manipulation;
+}
+
+button:focus-visible,
+input:focus-visible,
+textarea:focus-visible,
+select:focus-visible {
+  outline: 2px solid var(--lime);
+  outline-offset: 2px;
+}
+
+::selection {
+  color: var(--ink);
+  background: var(--lime);
+}
+
+::-webkit-scrollbar {
+  width: 9px;
+  height: 9px;
+}
+
+::-webkit-scrollbar-track {
+  background: var(--ink);
+}
+
+::-webkit-scrollbar-thumb {
+  background: #30332b;
+  border: 2px solid var(--ink);
+}
+
+.fatal-screen {
+  min-height: 100vh;
+  display: grid;
+  place-content: center;
+  justify-items: start;
+  gap: 18px;
+  padding: clamp(28px, 8vw, 110px);
+  background:
+    linear-gradient(90deg, rgba(255,255,255,.025) 1px, transparent 1px) 0 0 / 52px 52px,
+    linear-gradient(rgba(255,255,255,.025) 1px, transparent 1px) 0 0 / 52px 52px,
+    radial-gradient(circle at 65% 35%, var(--coral-soft), transparent 32%),
+    var(--ink);
+}
+
+.fatal-mark {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  color: var(--coral);
+}
+
+.fatal-mark i {
+  width: 110px;
+  height: 1px;
+  background: linear-gradient(90deg, var(--coral), transparent);
+}
+
+.fatal-screen h1 {
+  max-width: 900px;
+  margin: 0;
+  color: var(--paper);
+  font-family: var(--display);
+  font-size: clamp(42px, 7vw, 92px);
+  font-weight: 400;
+  line-height: .9;
+  letter-spacing: -.055em;
+}
+
+.fatal-screen h1 em {
+  color: var(--coral);
+  font-weight: 300;
+}
+
+.fatal-screen p {
+  max-width: 620px;
+  margin: 0 0 12px;
+  color: var(--muted);
+  font-size: var(--type-control);
+  line-height: 1.8;
+}
+
+.fatal-screen .launch-button {
+  width: min(100%, 310px);
+}
