+++ src/styles/changes.css
@@ -1,0 +1,223 @@
+/* Changes room: workspace toolbar, git summary, file list, diff viewer. */
+
+/* Changes workbench */
+.workspace-toolbar {
+  min-height: 64px;
+  padding: 10px 12px 10px 18px;
+  display: flex;
+  align-items: flex-end;
+  gap: 10px;
+  border: 1px solid var(--line);
+  background: rgba(13,15,12,.75);
+}
+
+.workspace-toolbar label {
+  flex: 1;
+  display: grid;
+  gap: 6px;
+}
+
+.workspace-toolbar .icon-button {
+  width: 43px;
+  height: 43px;
+}
+
+.workspace-live-state {
+  height: 43px;
+  padding: 0 12px;
+  display: inline-flex;
+  align-items: center;
+  gap: 8px;
+  border: 1px solid var(--line);
+  color: var(--coral);
+  font-size: var(--type-meta);
+  letter-spacing: .09em;
+  white-space: nowrap;
+}
+
+.workspace-live-state.is-connected {
+  color: var(--paper);
+}
+
+.git-summary-strip {
+  margin-top: 12px;
+  display: grid;
+  grid-template-columns: 1.2fr repeat(4, .7fr);
+  border: 1px solid var(--line);
+}
+
+.git-summary-strip > div {
+  min-height: 78px;
+  padding: 13px 16px;
+  display: grid;
+  grid-template-columns: 18px 1fr;
+  align-content: center;
+  column-gap: 8px;
+  border-right: 1px solid var(--line);
+}
+
+.git-summary-strip > div:last-child { border-right: 0; }
+
+.git-summary-strip svg {
+  grid-row: 1 / 3;
+  align-self: center;
+  color: var(--muted);
+}
+
+.git-summary-strip span {
+  color: var(--muted);
+  font-size: var(--type-label);
+  letter-spacing: .14em;
+}
+
+.git-summary-strip strong {
+  margin-top: 5px;
+  overflow: hidden;
+  font-family: 'Syne Variable', sans-serif;
+  font-size: 17px;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.git-summary-strip .git-add strong,
+.git-summary-strip .git-add svg { color: var(--lime); }
+
+.git-summary-strip .git-delete strong,
+.git-summary-strip .git-delete svg { color: var(--coral); }
+
+.changes-workbench {
+  height: min(650px, calc(100vh - 360px));
+  min-height: 480px;
+  display: grid;
+  grid-template-columns: minmax(270px, .34fr) minmax(0, .66fr);
+  gap: 12px;
+}
+
+.change-files,
+.diff-viewer {
+  min-height: 0;
+  overflow: hidden;
+}
+
+.change-files > header,
+.diff-viewer > header {
+  height: 62px;
+  padding: 18px;
+  border-bottom: 1px solid var(--line);
+}
+
+.change-files > header > span,
+.diff-truncated {
+  color: var(--muted);
+  font-size: var(--type-meta);
+  letter-spacing: .12em;
+}
+
+.file-filter {
+  height: 44px;
+  padding: 0 13px;
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  color: var(--muted);
+  border-bottom: 1px solid var(--line);
+}
+
+.file-filter input {
+  min-width: 0;
+  flex: 1;
+  color: var(--paper);
+  border: 0;
+  background: transparent;
+  outline: 0;
+  font-size: var(--type-body);
+}
+
+.change-file-list {
+  height: calc(100% - 106px);
+  overflow: auto;
+}
+
+.change-file-list button {
+  width: 100%;
+  min-height: 52px;
+  padding: 8px 11px;
+  display: grid;
+  grid-template-columns: 25px minmax(0, 1fr) auto;
+  align-items: center;
+  gap: 8px;
+  color: var(--paper);
+  border: 0;
+  border-bottom: 1px solid var(--line);
+  background: transparent;
+  text-align: left;
+  cursor: pointer;
+}
+
+.change-file-list button:hover,
+.change-file-list button.is-active { background: var(--lime-soft); }
+.change-file-list button.is-active { box-shadow: inset 2px 0 var(--lime); }
+
+.file-status {
+  color: var(--amber);
+  font-size: var(--type-meta);
+  text-align: center;
+}
+
+.file-path {
+  overflow: hidden;
+  font-size: var(--type-meta);
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.file-delta {
+  display: flex;
+  gap: 6px;
+  font-size: var(--type-label);
+}
+
+.file-delta i { color: var(--lime); font-style: normal; }
+.file-delta em { color: var(--coral); font-style: normal; }
+
+.files-empty {
+  padding: 40px 20px;
+  color: var(--muted);
+  text-align: center;
+  font-size: var(--type-body);
+}
+
+.diff-viewer {
+  display: flex;
+  flex-direction: column;
+}
+
+.diff-viewer h2 {
+  max-width: 620px;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.diff-viewer pre {
+  flex: 1;
+  min-height: 0;
+  margin: 0;
+  padding: 18px 20px 80px;
+  overflow: auto;
+  color: #cfd1ca;
+  background: linear-gradient(90deg, rgba(217,255,67,.025) 1px, transparent 1px) 0 0 / 42px 100%, #090a08;
+  font: var(--type-body)/1.7 'IBM Plex Mono', monospace;
+  white-space: pre;
+  tab-size: 2;
+}
+
+.diff-empty {
+  flex: 1;
+  display: grid;
+  place-content: center;
+  justify-items: center;
+  gap: 12px;
+  color: var(--muted);
+  font-size: var(--type-body);
+}
