+++ src/styles/overview.css
@@ -1,0 +1,308 @@
+/* Overview room: system card, metric cards. */
+
+.overview-grid {
+  display: grid;
+  grid-template-columns: minmax(400px, 1.45fr) repeat(2, minmax(180px, .6fr));
+  grid-template-rows: repeat(2, minmax(190px, auto));
+  gap: 10px;
+  margin-top: 26px;
+}
+
+.system-card {
+  position: relative;
+  grid-row: 1 / 3;
+  min-height: 390px;
+  overflow: hidden;
+  padding: 24px;
+  border: 1px solid rgba(217,255,67,.34);
+  background:
+    radial-gradient(circle at 70% 45%, rgba(217,255,67,.09), transparent 29%),
+    linear-gradient(145deg, #11140d 0%, #0c0e0a 100%);
+}
+
+.panel-cut {
+  clip-path: polygon(0 0, calc(100% - 26px) 0, 100% 26px, 100% 100%, 0 100%);
+}
+
+.system-card::after {
+  content: '';
+  position: absolute;
+  width: 52px;
+  height: 52px;
+  right: -27px;
+  top: -27px;
+  border: 1px solid var(--lime);
+  transform: rotate(45deg);
+}
+
+.system-card-grid {
+  position: absolute;
+  inset: 0;
+  opacity: .16;
+  background:
+    linear-gradient(90deg, var(--lime) 1px, transparent 1px) 0 0 / 45px 45px,
+    linear-gradient(var(--lime) 1px, transparent 1px) 0 0 / 45px 45px;
+  mask-image: linear-gradient(90deg, transparent 35%, black);
+}
+
+.system-card-copy {
+  position: relative;
+  z-index: 2;
+  padding-top: 54px;
+}
+
+.hero-number {
+  margin: 7px 0 -6px;
+  color: var(--paper);
+  font-size: clamp(78px, 9vw, 132px);
+  line-height: .9;
+  letter-spacing: -.08em;
+  font-weight: 720;
+}
+
+.hero-label {
+  color: var(--lime);
+  font-size: var(--type-control);
+  letter-spacing: .2em;
+}
+
+.system-readouts {
+  width: 54%;
+  min-width: 290px;
+  margin-top: 38px;
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  border-top: 1px solid var(--line);
+  border-bottom: 1px solid var(--line);
+}
+
+.system-readouts > div {
+  display: grid;
+  gap: 7px;
+  padding: 12px 13px 11px 0;
+  border-right: 1px solid var(--line);
+}
+
+.system-readouts > div + div {
+  padding-left: 13px;
+}
+
+.system-readouts > div:last-child {
+  border-right: 0;
+}
+
+.system-readouts span {
+  color: var(--muted);
+  font-size: var(--type-meta);
+  text-transform: uppercase;
+}
+
+.system-readouts strong {
+  font-family: 'Syne Variable', sans-serif;
+  font-size: 18px;
+}
+
+.system-orbit {
+  position: absolute;
+  z-index: 1;
+  width: min(27vw, 320px);
+  aspect-ratio: 1;
+  right: 2%;
+  top: 13%;
+}
+
+.orbit {
+  position: absolute;
+  border: 1px solid rgba(217,255,67,.3);
+  border-radius: 50%;
+}
+
+.orbit-one {
+  inset: 5%;
+  animation: orbit-spin 18s linear infinite;
+}
+
+.orbit-two {
+  inset: 21%;
+  border-style: dashed;
+  animation: orbit-spin 11s linear infinite reverse;
+}
+
+.orbit-one::before,
+.orbit-two::before {
+  content: '';
+  position: absolute;
+  width: 5px;
+  height: 5px;
+  border-radius: 50%;
+  top: -3px;
+  left: 50%;
+  background: var(--lime);
+  box-shadow: 0 0 10px var(--lime);
+}
+
+@keyframes orbit-spin {
+  to { transform: rotate(360deg); }
+}
+
+.orbit-center {
+  position: absolute;
+  inset: 34%;
+  display: grid;
+  place-content: center;
+  text-align: center;
+  border: 1px solid var(--lime);
+  background: #0d100a;
+  transform: rotate(45deg);
+}
+
+.orbit-center span,
+.orbit-center small {
+  transform: rotate(-45deg);
+}
+
+.orbit-center span {
+  color: var(--lime);
+  font-family: 'Syne Variable', sans-serif;
+  font-size: 29px;
+  font-weight: 600;
+}
+
+.orbit-center small {
+  margin-top: -3px;
+  color: var(--muted);
+  font-size: var(--type-label);
+  letter-spacing: .1em;
+}
+
+.orbit-node {
+  position: absolute;
+  width: 4px;
+  height: 4px;
+  background: var(--paper);
+}
+
+.node-one { left: 9%; top: 49%; }
+.node-two { right: 12%; top: 28%; }
+.node-three { right: 26%; bottom: 10%; background: var(--coral); }
+
+.system-footer {
+  position: absolute;
+  z-index: 3;
+  left: 24px;
+  right: 24px;
+  bottom: 19px;
+  display: flex;
+  justify-content: space-between;
+  gap: 20px;
+  color: var(--muted-dim);
+  font-size: var(--type-meta);
+  letter-spacing: .08em;
+  text-transform: uppercase;
+}
+
+.system-footer > span {
+  min-width: 0;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.system-footer > span:first-child {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  flex: 0 0 auto;
+  color: var(--paper);
+}
+
+.metric-card {
+  position: relative;
+  min-height: 190px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  overflow: hidden;
+  padding: 19px 20px 18px;
+  border: 1px solid var(--line);
+  background: rgba(16,18,14,.78);
+  transition: transform .2s ease, border-color .2s ease;
+}
+
+.metric-card:hover {
+  transform: translateY(-2px);
+  border-color: var(--line-strong);
+}
+
+.metric-card::after {
+  content: '';
+  position: absolute;
+  width: 42px;
+  height: 42px;
+  right: -25px;
+  bottom: -25px;
+  transform: rotate(45deg);
+  background: currentColor;
+  opacity: .16;
+}
+
+.metric-card-top {
+  display: flex;
+  justify-content: space-between;
+  color: var(--muted-dim);
+  font-size: var(--type-body);
+}
+
+.tone-lime .metric-card-top svg,
+.tone-lime .metric-value {
+  color: var(--lime);
+}
+
+.tone-coral .metric-card-top svg,
+.tone-coral .metric-value {
+  color: var(--coral);
+}
+
+.metric-label {
+  margin-bottom: 6px;
+  color: var(--muted);
+  font-size: var(--type-body);
+  letter-spacing: .1em;
+  text-transform: uppercase;
+}
+
+.metric-value {
+  font-size: clamp(27px, 3vw, 42px);
+  line-height: 1;
+  letter-spacing: -.05em;
+}
+
+.metric-detail {
+  margin-top: 8px;
+  color: var(--muted-dim);
+  font-size: var(--type-body);
+}
+
+.metric-rule {
+  position: absolute;
+  left: 20px;
+  right: 20px;
+  bottom: 0;
+  height: 2px;
+  background: rgba(255,255,255,.035);
+}
+
+.metric-rule span {
+  display: block;
+  width: 38%;
+  height: 100%;
+  background: currentColor;
+}
+
+.tone-lime .metric-rule { color: var(--lime); }
+.tone-coral .metric-rule { color: var(--coral); }
+.tone-paper .metric-rule { color: var(--muted-dim); }
+
+.section-gap {
+  margin-top: 10px;
+}
