+++ src/styles/panels.css
@@ -1,0 +1,419 @@
+/* Shared panels, charts, ranked bars, session list, status glyphs, model mix, workspace list. */
+
+.two-col-grid {
+  display: grid;
+  grid-template-columns: minmax(0, 1.25fr) minmax(330px, .75fr);
+  gap: 10px;
+}
+
+.panel {
+  min-width: 0;
+  border: 1px solid var(--line);
+  background: rgba(13,15,12,.78);
+}
+
+.panel-head {
+  min-height: 60px;
+  padding: 0 18px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 12px;
+  border-bottom: 1px solid var(--line);
+}
+
+.panel-head > div {
+  display: flex;
+  align-items: center;
+  min-width: 0;
+  gap: 12px;
+}
+
+.panel-head h2 {
+  margin: 0;
+  font-family: 'Syne Variable', sans-serif;
+  font-size: 13px;
+  font-weight: 620;
+  letter-spacing: .01em;
+}
+
+.panel-meta {
+  color: var(--muted-dim);
+  font-size: var(--type-meta);
+  text-transform: uppercase;
+  letter-spacing: .08em;
+}
+
+.text-button {
+  display: flex;
+  align-items: center;
+  gap: 6px;
+  padding: 5px 0;
+  color: var(--muted);
+  border: 0;
+  background: none;
+  font-size: var(--type-meta);
+  text-transform: uppercase;
+  cursor: pointer;
+}
+
+.text-button:hover {
+  color: var(--lime);
+}
+
+.panel-body {
+  padding: 20px;
+}
+
+.chart-summary {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  border: 1px solid var(--line);
+}
+
+.chart-summary > div {
+  padding: 11px;
+  display: grid;
+  gap: 3px;
+  border-right: 1px solid var(--line);
+}
+
+.chart-summary > div:last-child {
+  border-right: 0;
+}
+
+.chart-summary strong {
+  font-family: 'Syne Variable', sans-serif;
+  font-size: 19px;
+}
+
+.chart-summary span {
+  color: var(--muted);
+  font-size: var(--type-label);
+  letter-spacing: .08em;
+  text-transform: uppercase;
+}
+
+.bar-chart {
+  height: 176px;
+  margin-top: 18px;
+  display: grid;
+  grid-template-columns: repeat(14, minmax(12px, 1fr));
+  align-items: stretch;
+  gap: 7px;
+  border-bottom: 1px solid var(--line);
+  background: repeating-linear-gradient(to bottom, var(--line) 0 1px, transparent 1px 44px);
+}
+
+.bar-column {
+  min-width: 0;
+  display: grid;
+  grid-template-rows: 1fr 20px;
+  align-items: end;
+  gap: 5px;
+  color: var(--muted-dim);
+  font-size: var(--type-label);
+  text-align: center;
+  text-transform: uppercase;
+}
+
+.bar-track {
+  position: relative;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: end;
+}
+
+.bar-fill {
+  width: 100%;
+  min-height: 3px;
+  background: linear-gradient(to top, #7c931e, var(--lime));
+  transform-origin: bottom;
+  animation: bars-up .6s cubic-bezier(.22,.8,.22,1) both;
+}
+
+.bar-fill.has-error {
+  box-shadow: inset 0 3px var(--coral);
+}
+
+@keyframes bars-up {
+  from { transform: scaleY(0); }
+}
+
+.chart-legend {
+  margin-top: 12px;
+  display: flex;
+  gap: 18px;
+  color: var(--muted-dim);
+  font-size: var(--type-label);
+  text-transform: uppercase;
+}
+
+.chart-legend span {
+  display: flex;
+  align-items: center;
+  gap: 6px;
+}
+
+.chart-legend i {
+  width: 14px;
+  height: 2px;
+}
+
+.legend-lime { background: var(--lime); }
+.legend-coral { background: var(--coral); }
+
+.ranked-bars {
+  display: grid;
+  gap: 14px;
+}
+
+.rank-row {
+  display: grid;
+  grid-template-columns: 22px 1fr;
+  gap: 9px;
+  align-items: end;
+}
+
+.rank-number {
+  color: var(--muted-dim);
+  font-size: var(--type-meta);
+}
+
+.rank-main {
+  display: grid;
+  gap: 7px;
+}
+
+.rank-copy {
+  display: flex;
+  justify-content: space-between;
+  min-width: 0;
+  font-size: var(--type-body);
+}
+
+.rank-copy strong {
+  max-width: 85%;
+  overflow: hidden;
+  font-weight: 450;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.rank-copy span {
+  color: var(--muted);
+}
+
+.rank-track {
+  height: 2px;
+  background: rgba(255,255,255,.06);
+}
+
+.rank-track span {
+  display: block;
+  height: 100%;
+  min-width: 3px;
+  background: var(--lime);
+  box-shadow: 0 0 7px rgba(217,255,67,.25);
+}
+
+.lower-grid {
+  grid-template-columns: minmax(0, 1.4fr) minmax(330px, .6fr);
+}
+
+.side-stack {
+  display: grid;
+  gap: 10px;
+}
+
+.session-list {
+  margin: -20px;
+}
+
+.session-row {
+  width: 100%;
+  min-height: 65px;
+  padding: 10px 14px 10px 18px;
+  display: grid;
+  grid-template-columns: 18px minmax(0, 1fr) auto 60px 17px;
+  align-items: center;
+  gap: 10px;
+  color: var(--paper);
+  border: 0;
+  border-bottom: 1px solid var(--line);
+  background: transparent;
+  text-align: left;
+  cursor: pointer;
+  transition: background .15s ease;
+}
+
+.session-row:last-child {
+  border-bottom: 0;
+}
+
+.session-row:hover {
+  background: rgba(217,255,67,.035);
+}
+
+.session-copy {
+  display: grid;
+  min-width: 0;
+  gap: 4px;
+}
+
+.session-copy strong {
+  overflow: hidden;
+  font-family: 'Syne Variable', sans-serif;
+  font-size: 11px;
+  font-weight: 560;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.session-copy span,
+.session-metrics,
+.session-row time {
+  color: var(--muted-dim);
+  font-size: var(--type-meta);
+}
+
+.session-metrics {
+  display: flex;
+  gap: 13px;
+}
+
+.session-row time {
+  text-align: right;
+}
+
+.status-glyph {
+  position: relative;
+  width: 13px;
+  height: 13px;
+  display: inline-grid;
+  place-items: center;
+  flex: 0 0 auto;
+  border: 1px solid #4e524a;
+  border-radius: 50%;
+}
+
+.status-glyph > span {
+  width: 3px;
+  height: 3px;
+  border-radius: 50%;
+  background: #575b52;
+}
+
+.status-live,
+.status-recent {
+  border-color: rgba(217,255,67,.65);
+}
+
+.status-live > span,
+.status-recent > span {
+  background: var(--lime);
+}
+
+.status-live::after {
+  content: '';
+  position: absolute;
+  inset: -4px;
+  border: 1px solid rgba(217,255,67,.2);
+  border-radius: 50%;
+  animation: status-ring 1.7s ease-out infinite;
+}
+
+@keyframes status-ring {
+  to { transform: scale(1.35); opacity: 0; }
+}
+
+.status-attention {
+  border-color: var(--coral);
+}
+
+.status-attention > span {
+  background: var(--coral);
+}
+
+.segmented-track {
+  height: 9px;
+  display: flex;
+  gap: 2px;
+}
+
+.segment {
+  display: block;
+  min-width: 3px;
+}
+
+.segment-0 { background-color: var(--lime) !important; }
+.segment-1 { background-color: #a5bd3e !important; }
+.segment-2 { background-color: #7a8171 !important; }
+.segment-3 { background-color: #4c5048 !important; }
+
+.model-list {
+  margin-top: 16px;
+  display: grid;
+  gap: 9px;
+}
+
+.model-list > div {
+  display: grid;
+  grid-template-columns: 7px 1fr auto;
+  align-items: center;
+  gap: 8px;
+  font-size: var(--type-meta);
+}
+
+.model-dot {
+  width: 5px;
+  height: 5px;
+}
+
+.model-list span {
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.model-list strong {
+  color: var(--muted);
+  font-weight: 400;
+}
+
+.workspace-list {
+  display: grid;
+  gap: 1px;
+}
+
+.workspace-list > div {
+  min-height: 34px;
+  display: grid;
+  grid-template-columns: 23px 18px 1fr auto;
+  align-items: center;
+  gap: 5px;
+  border-bottom: 1px solid rgba(255,255,255,.055);
+  font-size: var(--type-meta);
+}
+
+.workspace-list > div:last-child {
+  border-bottom: 0;
+}
+
+.workspace-index,
+.workspace-list > div > span:last-child {
+  color: var(--muted-dim);
+}
+
+.workspace-list strong {
+  min-width: 0;
+  overflow: hidden;
+  font-weight: 450;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.workspace-list svg {
+  color: var(--lime);
+}
